+++ app/(tabs)/_layout.tsx
@@ -20,26 +20,38 @@ export default function TabLayout() {
         tabBarBackground: TabBarBackground,
         tabBarStyle: Platform.select({
           ios: {
-            // Use a transparent background on iOS to show the blur effect
             position: 'absolute',
           },
           default: {},
         }),
-      }}>
+      }}
+    >
       <Tabs.Screen
         name="index"
         options={{
           title: 'Home',
-          tabBarIcon: ({ color }) => <IconSymbol size={28} name="house.fill" color={color} />,
-        }}
-      />
+          tabBarIcon: ({ color }) => (
+            <IconSymbol size={28} name="house.fill" color={color} />
+          ),
+        }} />
       <Tabs.Screen
-        name="welcomePage"
+        name="registrationScreen"
         options={{
-          title: 'Welcome Page',
-          tabBarIcon: ({ color }) => <IconSymbol size={28} name="house.fill" color={color} />,
-        }}
-      />
+          title: 'Register',
+          tabBarStyle: { display: 'none' },
+        }} />
+      <Tabs.Screen
+        name="registrationAttendanceConfirmation"
+        options={{
+          title: 'registrationAttendanceConfirmation',
+          tabBarStyle: { display: 'none' },
+        }} />
+        <Tabs.Screen
+        name="registrationTransportationConfirmation"
+        options={{
+          title: 'registrationTransportationConfirmation',
+          tabBarStyle: { display: 'none' },
+        }} />
     </Tabs>
   );
 }
