+++ app/(tabs)/index.tsx
@@ -1,75 +1,126 @@
-import { Image } from 'expo-image';
-import { Platform, StyleSheet } from 'react-native';
+import WelcomePageBackground from '@/assets/images/svg/welcomePageBackground';
+import en from "@/assets/translations/en.json";
+import { router } from 'expo-router';
+import React from 'react';
+import { StyleSheet, Text, TouchableOpacity, View } from 'react-native';
 
-import { HelloWave } from '@/components/HelloWave';
-import ParallaxScrollView from '@/components/ParallaxScrollView';
-import { ThemedText } from '@/components/ThemedText';
-import { ThemedView } from '@/components/ThemedView';
 
-export default function HomeScreen() {
+
+export default function homeScreen(){
   return (
-    <ParallaxScrollView
-      headerBackgroundColor={{ light: '#A1CEDC', dark: '#1D3D47' }}
-      headerImage={
-        <Image
-          source={require('@/assets/images/partial-react-logo.png')}
-          style={styles.reactLogo}
-        />
-      }>
-      <ThemedView style={styles.titleContainer}>
-        <ThemedText type="title">Welcome!</ThemedText>
-        <HelloWave />
-      </ThemedView>
-      <ThemedView style={styles.stepContainer}>
-        <ThemedText type="subtitle">Step 1: Try it</ThemedText>
-        <ThemedText>
-          Edit <ThemedText type="defaultSemiBold">app/(tabs)/index.tsx</ThemedText> to see changes.
-          Press{' '}
-          <ThemedText type="defaultSemiBold">
-            {Platform.select({
-              ios: 'cmd + d',
-              android: 'cmd + m',
-              web: 'F12',
-            })}
-          </ThemedText>{' '}
-          to open developer tools.
-        </ThemedText>
-      </ThemedView>
-      <ThemedView style={styles.stepContainer}>
-        <ThemedText type="subtitle">Step 2: Explore</ThemedText>
-        <ThemedText>
-          {`Tap the Explore tab to learn more about what's included in this starter app.`}
-        </ThemedText>
-      </ThemedView>
-      <ThemedView style={styles.stepContainer}>
-        <ThemedText type="subtitle">Step 3: Get a fresh start</ThemedText>
-        <ThemedText>
-          {`When you're ready, run `}
-          <ThemedText type="defaultSemiBold">npm run reset-project</ThemedText> to get a fresh{' '}
-          <ThemedText type="defaultSemiBold">app</ThemedText> directory. This will move the current{' '}
-          <ThemedText type="defaultSemiBold">app</ThemedText> to{' '}
-          <ThemedText type="defaultSemiBold">app-example</ThemedText>.
-        </ThemedText>
-      </ThemedView>
-    </ParallaxScrollView>
+    <View style={styles.container}>
+      <WelcomePageBackground />
+      <View style={
+        { flex: 1, justifyContent: 'center', alignItems: 'center' }
+      } >
+        <Text style={styles.title}>{en.africaConsulting}</Text>
+        <Text style={styles.subtitle}>{en.welcomePage.subtitle}</Text>
+        <View style={styles.card}>
+          <Text style={styles.builtBy}>{en.welcomePage.builtBy}</Text>
+          <Text style={styles.cardTitle}>{en.welcomePage.cardTitle}</Text>
+          <Text style={styles.cardSubtitle}>{en.welcomePage.cardSubtitle}</Text>
+          <View style={styles.divider} />
+          <Text style={styles.hostedBy}>{en.welcomePage.hostedBy}</Text>
+          <Text style={styles.cardSubtitle}>{en.welcomePage.cardSubtitlePractice}</Text>
+        </View>
+        <TouchableOpacity style={styles.button} onPress={ () => router.push('/registrationScreen')}>
+          <Text style={styles.buttonText}>{en.welcomePage.buttonText}</Text>
+        </TouchableOpacity>
+      </View>
+    </View>
   );
 }
 
 const styles = StyleSheet.create({
-  titleContainer: {
-    flexDirection: 'row',
+  container: {
+    flex: 1,
+    backgroundColor: '#000',
     alignItems: 'center',
-    gap: 8,
+    paddingTop: 60,
+    paddingHorizontal: 16,
   },
-  stepContainer: {
-    gap: 8,
-    marginBottom: 8,
+  headerImage: {
+    width: 220,
+    height: 180,
+    marginBottom: 32,
   },
-  reactLogo: {
-    height: 178,
-    width: 290,
-    bottom: 0,
-    left: 0,
+  title: {
+    color: '#fff',
+    fontSize: 24,
+    fontWeight: 'bold',
+    textAlign: 'center',
+  },
+  subtitle: {
+    color: '#fff',
+    fontSize: 18,
+    marginTop: 8,
+    marginBottom: 32,
+    textAlign: 'center',
+  },
+  card: {
+    backgroundColor: '#181818',
+    borderColor: '#444',
+    borderWidth: 1,
+    borderRadius: 8,
+    padding: 20,
+    width: '100%',
+    marginBottom: 40,
+  },
+  builtBy: {
+    color: '#B6E23A',
+    fontSize: 12,
+    fontWeight: 'bold',
+    marginBottom: 4,
+    textAlign: 'center',
+  },
+  hostedBy: {
+    color: '#B6E23A',
+    fontSize: 12,
+    fontWeight: 'bold',
+    marginTop: 12,
+    marginBottom: 4,
+    textAlign: 'center',
+  },
+  cardTitle: {
+    color: '#fff',
+    fontSize: 16,
+    fontWeight: 'bold',
+    textAlign: 'center',
+  },
+  cardSubtitle: {
+    color: '#ccc',
+    fontSize: 13,
+    textAlign: 'center',
+  },
+  divider: {
+    borderBottomColor: '#444',
+    borderBottomWidth: 1,
+    marginVertical: 10,
+  },
+  button: {
+    backgroundColor: '#B6E23A',
+    borderRadius: 24,
+    paddingVertical: 14,
+    paddingHorizontal: 40,
+    marginTop: 16,
+    marginBottom: 32,
+    width: 220,
+    alignItems: 'center',
+  },
+  buttonText: {
+    color: '#fff',
+    fontWeight: 'bold',
+    fontSize: 18,
+  },
+  buttonHover: {
+  backgroundColor: 'darkgreen',
+},
+  footerImage: {
     position: 'absolute',
+    bottom: 0,
+    width: '120%',
+    height: 120,
+    left: '-10%',
+    opacity: 0.5,
   },
 });
