+++ app/(tabs)/registrationTransportationConfirmation.tsx
@@ -1,14 +1,21 @@
-import RegistrationScreen1Background from '@/assets/images/svg/registrationScreen1Background';
+import RegistrationScreenBackground from '@/assets/images/svg/registrationScreenBackground';
+import { router } from 'expo-router';
 import React, { useState } from 'react';
-import { KeyboardAvoidingView, Platform, StyleSheet, Text, TextInput, TouchableOpacity, View } from 'react-native';
+import {
+  KeyboardAvoidingView,
+  Platform,
+  StyleSheet,
+  Text,
+  TextInput,
+  TouchableOpacity,
+  View,
+} from 'react-native';
 import Icon from 'react-native-vector-icons/Ionicons';
 
-
-
 const transportationOptions = [
-  { label: "Need airport pickup", value: 'pickup' },
-  { label: "Will arrange my own", value: 'own' },
-  { label: "Not decided yet", value: 'undecided' },
+  { label: 'Need airport pickup', value: 'pickup' },
+  { label: 'Will arrange my own', value: 'own' },
+  { label: 'Not decided yet', value: 'undecided' },
 ];
 
 export default function RegistrationTransportationConfirmation() {
@@ -16,48 +23,72 @@ export default function RegistrationTransportationConfirmation() {
   const [message, setMessage] = useState('');
 
   return (
-    <View style={{ flex: 1, backgroundColor: '#000' }} >
-      <RegistrationScreen1Background />
+    <View style={{ flex: 1, backgroundColor: '#000' }}>
+      <RegistrationScreenBackground />
       <KeyboardAvoidingView
         style={{ flex: 1 }}
         behavior={Platform.OS === 'ios' ? 'padding' : undefined}
       >
-        <View style={[styles.overlay, { flex: 1, justifyContent: 'center', alignItems: 'center' }]}>
-          <Text style={styles.title}>Confirm your transportation</Text>
-          <Text style={styles.subtitle}>Let us know your preference</Text>
-          <View style={{ marginTop: 16, marginBottom: 18 }}>
-            {transportationOptions.map(option => (
-              <TouchableOpacity
-                key={option.value}
-                style={styles.radioRow}
-                onPress={() => setSelected(option.value)}
-                activeOpacity={0.7}
-              >
-                <View style={styles.radioOuter}>
-                  {selected === option.value && <View style={styles.radioInner} />}
-                </View>
-                <Text style={styles.radioLabel}>{option.label}</Text>
-              </TouchableOpacity>
-            ))}
+        <View style={{ flex: 1, justifyContent: 'center' }}>
+          {/* CARD */}
+          <View style={styles.overlay}>
+            <Text style={styles.title}>Confirm your transportation</Text>
+            <Text style={styles.subtitle}>Let us know your preference</Text>
+
+            <View style={{ marginTop: 16, marginBottom: 18 }}>
+              {transportationOptions.map((option) => (
+                <TouchableOpacity
+                  key={option.value}
+                  style={styles.radioRow}
+                  onPress={() => setSelected(option.value)}
+                  activeOpacity={0.7}
+                >
+                  <View style={styles.radioOuter}>
+                    {selected === option.value && <View style={styles.radioInner} />}
+                  </View>
+                  <Text style={styles.radioLabel}>{option.label}</Text>
+                </TouchableOpacity>
+              ))}
+            </View>
+
+            <Text style={styles.label}>Message</Text>
+            <TextInput
+              style={styles.input}
+              placeholder="Enter your message"
+              placeholderTextColor="#bdbdbd"
+              value={message}
+              onChangeText={setMessage}
+              multiline
+              numberOfLines={4}
+              textAlignVertical="top"
+            />
+
+            <TouchableOpacity style={styles.button}>
+              <Icon name="car-outline" size={20} color="#fff" style={{ marginRight: 8 }} />
+              <Text style={styles.buttonText}>Submit</Text>
+            </TouchableOpacity>
+
+            <Text style={styles.footerText}>
+              Having problems? <Text style={styles.contactText}>Contact us</Text>
+            </Text>
           </View>
-          <Text style={styles.label}>Message</Text>
-          <TextInput
-            style={styles.input}
-            placeholder="Enter your message"
-            placeholderTextColor="#bdbdbd"
-            value={message}
-            onChangeText={setMessage}
-            multiline
-            numberOfLines={4}
-            textAlignVertical="top"
-          />
-          <TouchableOpacity style={styles.button}>
-            <Icon name="car-outline" size={20} color="#fff" style={{ marginRight: 8 }} />
-            <Text style={styles.buttonText}>Submit</Text>
-          </TouchableOpacity>
-          <Text style={styles.footerText}>
-            Having problems? <Text style={styles.contactText}>Contact us</Text>
-          </Text>
+
+          {/* BUTTONS OUTSIDE CARD */}
+          <View style={styles.bottomButtonsContainer}>
+            <TouchableOpacity
+              style={styles.navButton}
+              onPress={() => router.back()}
+            >
+              <Text style={styles.backButtonText}>Back</Text>
+            </TouchableOpacity>
+
+            <TouchableOpacity
+              style={styles.navButton}
+              onPress={() => router.push('/(tabs)/home1')}
+            >
+              <Text style={styles.backButtonText}>Skip</Text>
+            </TouchableOpacity>
+          </View>
         </View>
       </KeyboardAvoidingView>
     </View>
@@ -65,18 +96,12 @@ export default function RegistrationTransportationConfirmation() {
 }
 
 const styles = StyleSheet.create({
-  bg: {
-    flex: 1,
-    width: '100%',
-    height: '100%',
-  },
   overlay: {
     backgroundColor: 'rgba(0,0,0,0.7)',
     borderRadius: 18,
     padding: 24,
-    margin: 24,
-    marginTop: 80,
-    alignItems: 'flex-start',
+    marginHorizontal: 24,
+    marginTop: 30,
     shadowColor: '#000',
     shadowOpacity: 0.25,
     shadowRadius: 12,
@@ -123,7 +148,6 @@ const styles = StyleSheet.create({
     color: '#fff',
     fontSize: 13,
     marginBottom: 4,
-    marginLeft: 2,
   },
   input: {
     backgroundColor: '#fff',
@@ -156,11 +180,29 @@ const styles = StyleSheet.create({
     color: '#fff',
     fontSize: 14,
     textAlign: 'center',
-    alignSelf: 'center',
     marginTop: 8,
   },
   contactText: {
     color: '#8DD22A',
     fontWeight: 'bold',
+  },
+  bottomButtonsContainer: {
+    flexDirection: 'row',
+    justifyContent: 'center',
+    marginTop: 16,
+    marginHorizontal: 24,
+  },
+  navButton: {
+    backgroundColor: '#000',
+    borderRadius: 20,
+    paddingVertical: 10,
+    paddingHorizontal: 32,
+    marginHorizontal: 8,
+    borderWidth: 1,
+    borderColor: '#fff',
+  },
+  backButtonText: {
+    color: '#fff',
+    fontSize: 16,
   },
 });
