+++ app/App.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import AppNavigator from './navigation/StackNavigator';
+
+export default function App() {
+  return <AppNavigator />;
+}
